v2.0: address AI review findings
- Rename to taskbar-auto-hide-after-flash (name now matches behavior)
- English-first README and settings with :zh-CN localization
- Use RegisterWindowMessage(SHELLHOOK) instead of hardcoded 0xC028/0xC029
- Use HSHELL_FLASH constant instead of 0x8006
- Use GetTickCount64() monotonic clock instead of FILETIME
- Remove Wh_ModAfterInit retry loop (Sleep on main thread + hooks never applied)
- Remove compatAutoHideSpeed setting (didn't do what it claimed)
- Remove enabled setting (redundant with Windhawk UI toggle)
- Add @architecture x86-64 (avoids x86 thiscall/stdcall mismatch crash)
- Clamp negative delay values with std::max(0, ...)
- Use LoadLibraryEx with LOAD_LIBRARY_SEARCH_SYSTEM32
- Qualify WindhawkUtils::HookSymbols
- Consistent g_ prefix on all globals
- Add credits to Cirn09's taskbar-autohide-better
- Wh_ModSettingsChanged void variant
